feat: support rendering matrix sample ancestor label and line spans across descendant sample labels
